Add a version of static methods
The implementation uses the idea as outlined in the challenge.
LoxClass extends LoxInstance. So each class can have an optional
metaclass. The parser adds methods that are preceded with a class
keyword to a list of metaclass methods. If there are metaclass methods
a metaclass is created and assigned to the class in when parsing the
class declaration. The interpreter then builds the class object and
adds the metaclass object if needed. The static methods are also added
binding them to the class (which in this case is the instance).

The resolver naturally also needs to resolve the static methods